# tools/swagger/docs/api-internal/iam-internal/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ces APIS sont utilisés pour configurer les clients, tenants, idp, groupes, profils, et utilisateurs.",
    "version": "v1",
    "title": "API IAM - Gestion des accès aux identités",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Direction de la diffusion et des partenariats",
      "email": "contact@programmevitam.fr"
    },
    "license": {
      "name": "License to be defined ...",
      "url": "http://notedefined.com"
    }
  },
  "host": "localhost",
  "basePath": "/iam/v1",
...
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/groups": {
      "get": {
        "tags": [
          "Groupes"
        ],
        "summary": "Récupération paginée des groupes",
        "operationId": "getAllPaginatedUsingGET_1",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          },
          {
            "name": "embedded",
            "in": "query",
            "description": "Récupération des profiles associés",
            "required": false,
            "type": "string",
            "enum": [
              "ALL"
            ]
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«GroupDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Groupes"
        ],
        "summary": "Création d'un groupe",
        "operationId": "createUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "groupe",
            "description": "Groupe à créer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          },
          "401": {
            "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
          },
          "403": {
            "description": "L'utilisateur ne possède pas les droits pour effectuer cette opération"
          }
        }
      }
    },
...
    "/subrogations/groups/{id}": {
      "get": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Récupération d'un groupe",
        "operationId": "getGroupByIdUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/subrogations/me/superuser": {
      "get": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Récupération de ma subrogation en tant que subrogateur",
        "operationId": "getMySubrogationAsSuperuserUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SubrogationDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/subrogations/me/surrogate": {
      "get": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Récupération de ma subrogation en tant que subrogé",
        "operationId": "getMySubrogationAsSurrogateUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/x-application-id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SubrogationDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/subrogations/surrogate/accept/{id}": {
      "patch": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Acceptation d'une subrogation par le subrogé",
        "operationId": "acceptUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique de la subrogation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SubrogationDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/subrogations/surrogate/decline/{id}": {
      "delete": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Refus de la subrogation par le subrogé",
        "operationId": "declineUsingDELETE",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique de la subrogation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/subrogations/users": {
      "get": {
        "tags": [
          "Subrogations"
        ],
        "summary": "Récupération paginée des utilisateurs",
        "operationId": "getUsersUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "criteria",
            "in": "query",
            "description": "criteria",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "orderBy",
            "required": false,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "description": "direction",
            "required": false,
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«UserDto»"
            }
          },
          "401": {
            "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
          },
          "404": {
            "description": "Non trouvée"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération paginée des utilisateurs",
        "operationId": "getAllPaginatedUsingGET_3",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "$ref": "#/parameters/page"
          },
          {
            "$ref": "#/parameters/size"
          },
          {
            "$ref": "#/parameters/criteria"
          },
          {
            "$ref": "#/parameters/orderBy"
          },
          {
            "$ref": "#/parameters/direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des utilisateurs",
            "schema": {
              "$ref": "#/definitions/PaginatedValuesDto«UserDto»"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Création d'un utilisateur",
        "operationId": "createUsingPOST_7",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "in": "body",
            "name": "user",
            "description": "Utilisateur à créer",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "L'utilisateur a été créé avec succès",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "responses": {
    "Unauthorized": {
      "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
    },
    "NotFound": {
      "description": "Ressource non trouvée"
    }
  },
  "parameters": {
    "page": {
      "in": "query",
      "name": "page",
      "required": true,
      "minimum": 0,
      "type": "integer",
      "format": "int32",
      "description": "Numéro de la page"
    },
    "size": {
      "name": "size",
      "in": "query",
      "required": true,
      "minimum": 1,
      "type": "integer",
      "format": "int32",
      "description": "Nombre de résultats maximum"
    },
    "criteria": {
      "name": "criteria",
      "in": "query",
      "description": "Critères de recherche",
      "required": false,
      "type": "string"
    },
    "orderBy": {
      "name": "orderBy",
      "in": "query",
      "description": "Nom du champ sur lequel effectuer le tri",
      "required": false,
      "type": "string"
    },
    "direction": {
      "name": "direction",
      "in": "query",
      "description": "Direction pour le tri",
      "required": false,
      "type": "string",
      "enum": [
        "ASC",
        "DESC"
      ]
    },
    "x-tenant-id": {
      "name": "X-Tenant-Id",
      "in": "header",
      "description": "Identifiant du tenant sur lequel doit être exécuté la requête",
      "required": true,
      "type": "integer",
      "format": "int32"
    },
    "x-user-token": {
      "name": "X-User-Token",
      "in": "header",
      "description": "Identifiant technique du token généré par le serveur CAS lors de l'authentification",
      "required": true,
      "type": "string"
    },
    "x-application-id": {
      "name": "X-Application-Id",
      "in": "header",
      "description": "Identifiant unique utilisé dans la génération des logs et des journaux VITAM",
      "required": false,
      "type": "string"
    },
    "x-user-level": {
      "name": "X-User-Level",
      "in": "header",
      "description": "Niveau de l'utilisateur",
      "required": true,
      "type": "string"
    },
    "x-customer-id": {
      "name": "X-Customer-Id",
      "in": "header",
      "description": "Identifiant technique du client",
      "required": true,
      "type": "string"
    }
  },
  "definitions": {
...
    "GroupDto": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique du client"
        },
        "description": {
          "type": "string",
          "example": "Groupe pour les utilisateurs RH"
        },
        "enabled": {
          "type": "boolean",
          "example": true,
          "description": "Actif ou Inactif"
        },
        "id": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique",
          "readOnly": true
        },
        "identifier": {
          "type": "string",
          "readOnly": true,
          "example": "1",
          "description": "Identifiant métier"
        },
        "level": {
          "type": "string",
          "example": "",
          "description": "Niveau"
        },
        "name": {
          "type": "string",
          "example": "Groupe RH",
          "description": "Nom"
        },
        "profileIds": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "1",
            "2",
            "3"
          ],
          "description": "Ensemble des identifiants technique des profils contenu dans le groupe"
        },
        "profiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProfileDto"
          },
          "description": "Ensemble des profiles du groupe",
          "readOnly": true
        },
        "readonly": {
          "type": "boolean",
          "description": "Lecture seule",
          "example": true,
          "readOnly": true
        },
        "usersCount": {
          "type": "integer",
          "format": "int64",
          "readOnly": true,
          "description": "Nombre d'utilisateurs contenu dans le groupe"
        }
      },
      "required": [
        "customerId",
        "name",
        "description",
        "profilleIds"
      ],
      "title": "GroupDto"
    },
...
    "PaginatedValuesDto«GroupDto»": {
      "type": "object",
      "properties": {
        "hasMore": {
          "type": "boolean"
        },
        "pageNum": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GroupDto"
          }
        }
      },
      "title": "PaginatedValuesDto«GroupDto»"
    },
...
    "PaginatedValuesDto«UserDto»": {
      "type": "object",
      "properties": {
        "hasMore": {
          "type": "boolean"
        },
        "pageNum": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDto"
          }
        }
      },
      "title": "PaginatedValuesDto«UserDto»"
    },
    "ProfileDto": {
      "type": "object",
      "properties": {
        "applicationName": {
          "type": "string",
          "description": "Identifiant de l'application métier",
          "example": "USERS_APP"
        },
        "customerId": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique du client"
        },
        "description": {
          "type": "string",
          "example": "profile des droits utilisateurs",
          "description": "Description"
        },
        "enabled": {
          "type": "boolean",
          "example": true,
          "description": "Actif ou Inactif"
        },
        "externalParamId": {
          "type": "string",
          "description": "Identifier technique d'un paramètre extérieur"
        },
        "externalParamIdentifier": {
          "type": "string",
          "description": "Identifier métier d'un paramètre extérieur"
        },
        "groupsCount": {
          "type": "integer",
          "format": "int64",
          "readOnly": true,
          "example": 2,
          "description": "Nombre de groupe contenant ce profile"
        },
        "id": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique",
          "readOnly": true
        },
        "identifier": {
          "type": "string",
          "readOnly": true,
          "example": "1",
          "description": "Identifiant métier"
        },
        "level": {
          "type": "string",
          "example": "",
          "description": "Niveau"
        },
        "name": {
          "type": "string",
          "description": "Nom",
          "example": "ProfileRH"
        },
        "readonly": {
          "type": "boolean",
          "description": "Lecture seule",
          "example": true,
          "readOnly": true
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          },
          "description": "Ensemble des rôles"
        },
        "tenantIdentifier": {
          "type": "integer",
          "format": "int32",
          "example": "1",
          "description": "Identifiant métier du tenant"
        },
        "tenantName": {
          "type": "string",
          "description": "Nom du tenant",
          "example": "Tenant 1",
          "readOnly": true
        },
        "usersCount": {
          "type": "integer",
          "format": "int64",
          "example": 10,
          "description": "Nombre d'utilisateurs",
          "readOnly": true
        }
      },
      "required": [
        "customerId",
        "name",
        "description",
        "applicationName",
        "tenantIdentifier"
      ],
      "title": "ProfileDto"
    },
    "Role": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "ROLE_CREATE_USERS",
          "description": "Nom du rôle"
        }
      },
      "required": [
        "name"
      ],
      "title": "Role"
    },
    "SubrogationDto": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time",
          "description": "Date de fin de la subrogation"
        },
        "id": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique",
          "readOnly": true
        },
        "status": {
          "type": "string",
          "enum": [
            "CREATED",
            "ACCEPTED"
          ],
          "description": "Statut",
          "example": "CREATED"
        },
        "superUser": {
          "type": "string",
          "description": "Email de l'utilisateur subrogateur",
          "example": "superuser@arhiveco.fr"
        },
        "superUserCustomerId": {
          "type": "string",
          "description": "Identifiant métier du client de l'utilisateur subrogateur",
          "example": "1"
        },
        "surrogate": {
          "type": "string",
          "description": "Email de l'utilisateur subrogée",
          "example": "surrogate@archiveco.fr"
        },
        "surrogateCustomerId": {
          "type": "string",
          "description": "Identifiant métier du client de l'utilisateur subrogée",
          "example": "1"
        }
      },
      "required": [
        "status",
        "surrogate",
        "superUser"
      ],
      "title": "SubrogationDto"
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique du client"
        },
        "email": {
          "type": "string",
          "example": "toto@example.fr",
          "description": "Adresse mail"
        },
        "firstname": {
          "type": "string",
          "example": "toto",
          "description": "Prénom"
        },
        "groupId": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique du groupe"
        },
        "id": {
          "type": "string",
          "example": "1",
          "description": "Identifiant technique",
          "readOnly": true
        },
        "identifier": {
          "type": "string",
          "readOnly": true,
          "example": "1",
          "description": "Identifiant métier"
        },
        "language": {
          "type": "string",
          "enum": [
            "FR",
            "DE",
            "EN"
          ],
          "description": "Langue"
        },
        "lastConnection": {
          "type": "string",
          "format": "date-time",
          "description": "Date de la derniére connexion de l'utilisateur au format ISO-8601"
        },
        "lastname": {
          "type": "string",
          "example": "DUPONT",
          "description": "Nom de famille"
        },
        "level": {
          "type": "string",
          "example": " ",
          "description": "Niveau"
        },
        "mobile": {
          "type": "string",
          "example": "+33667678250",
          "description": "Numéro de téléphone portable au format international"
        },
        "nbFailedAttempts": {
          "type": "integer",
          "format": "int32",
          "description": "Nombre de tantive échouée depuis la dernière connexion réussie"
        },
        "otp": {
          "type": "boolean",
          "description": "Double authentification activée ou non"
        },
        "passwordExpirationDate": {
          "type": "string",
          "format": "date-time",
          "description": "Date d'expiration du mot de passe"
        },
        "phone": {
          "type": "string",
          "example": "+33145868450",
          "description": "Numéro de téléphone fixe au format international"
        },
        "readonly": {
          "type": "boolean",
          "description": "Lecture seule",
          "example": true,
          "readOnly": true
        },
        "status": {
          "type": "string",
          "enum": [
            "ENABLED",
            "DISABLED",
            "BLOCKED",
            "ANONYM"
          ],
          "description": "Statut"
        },
        "subrogeable": {
          "type": "boolean",
          "description": "Subrogeable"
        },
        "type": {
          "type": "string",
          "enum": [
            "GENERIC",
            "NOMINATIVE"
          ],
          "description": "Type"
        }
      },
      "required": [
        "customerId",
        "lastname",
        "firstname",
        "groupId",
        "language",
        "email"
      ],
      "title": "UserDto"
    }
  }
}